Move extended stabilizer code to specific file
